# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parser2 API Documentation",
        "description": "Advanced parsing, processing, normalization, and categorization of data from multiple sources",
        "contact": {
            "email": "support@parser2.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://dev.parser2.local",
            "description": "Parser2 API Server"
        }
    ],
    "paths": {
        "/api/health": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Check API health status",
                "description": "Returns the current health status of the API",
                "operationId": "healthCheck",
                "responses": {
                    "200": {
                        "description": "API is healthy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-01T12:00:00Z"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0.0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/parsers/parse": {
            "post": {
                "tags": [
                    "Parsers"
                ],
                "summary": "Execute a parse request",
                "description": "Parse content from a source using the specified parser type",
                "operationId": "fbe2ce803e0032e37f167edd9e0f902f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "source",
                                    "type"
                                ],
                                "properties": {
                                    "source": {
                                        "description": "Source URL or identifier to parse",
                                        "type": "string",
                                        "example": "https://reddit.com/r/programming.json"
                                    },
                                    "type": {
                                        "description": "Parser type to use",
                                        "type": "string",
                                        "enum": [
                                            "feeds",
                                            "reddit",
                                            "single_page",
                                            "telegram",
                                            "medium",
                                            "bing",
                                            "multi",
                                            "craigslist"
                                        ],
                                        "example": "reddit"
                                    },
                                    "keywords": {
                                        "description": "Optional keywords for search-based parsers",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "php",
                                            "laravel"
                                        ]
                                    },
                                    "options": {
                                        "description": "Parser-specific options",
                                        "type": "object"
                                    },
                                    "limit": {
                                        "description": "Maximum number of items to return",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "offset": {
                                        "description": "Offset for pagination",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "filters": {
                                        "description": "Additional filters",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parse result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "metadata": {
                                                    "type": "object"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "next_offset": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "429": {
                        "description": "Rate limit exceeded"
                    }
                }
            }
        },
        "/parsers": {
            "get": {
                "tags": [
                    "Parsers"
                ],
                "summary": "List available parsers",
                "description": "Get a list of all available parser types and their descriptions",
                "operationId": "18e91e4732410f40400a1b85be129ff7",
                "responses": {
                    "200": {
                        "description": "List of available parsers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "reddit"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Parser for reddit sources"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/parsers/{name}": {
            "get": {
                "tags": [
                    "Parsers"
                ],
                "summary": "Get parser details",
                "description": "Get detailed information about a specific parser",
                "operationId": "58c0ce89cdfb222acc1ea2b76b6ae58e",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Parser name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "feeds",
                                "reddit",
                                "single_page",
                                "telegram",
                                "medium",
                                "bing",
                                "multi",
                                "craigslist"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parser details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "reddit"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Parser for reddit sources"
                                                },
                                                "config": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Parser not found"
                    }
                }
            }
        },
        "/parsers/batch": {
            "post": {
                "tags": [
                    "Parsers"
                ],
                "summary": "Execute batch parse requests",
                "description": "Parse multiple sources in a single batch request (max 100)",
                "operationId": "6e136587b40a31857fa26cd7b600377e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "requests"
                                ],
                                "properties": {
                                    "requests": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "source": {
                                                    "type": "string",
                                                    "example": "https://example.com"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "single_page"
                                                },
                                                "keywords": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "options": {
                                                    "type": "object"
                                                },
                                                "limit": {
                                                    "type": "integer"
                                                },
                                                "offset": {
                                                    "type": "integer"
                                                },
                                                "filters": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Batch parse results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "results": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "summary": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "successful": {
                                                            "type": "integer",
                                                            "example": 8
                                                        },
                                                        "failed": {
                                                            "type": "integer",
                                                            "example": 2
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "System",
            "description": "System health and status endpoints"
        },
        {
            "name": "Parsers",
            "description": "Parser management and execution endpoints"
        }
    ]
}